+++ src/main/java/com/verdantartifice/primalmagick/common/loot/modifiers/LootModifierSerializersPM.java
@@ -29,4 +29,5 @@ public class LootModifierSerializersPM {
     public static final RegistryObject<Codec<RelicFragmentsModifier>> RELIC_FRAGMENTS = SERIALIZERS.register("relic_fragments", () -> RelicFragmentsModifier.CODEC);
     public static final RegistryObject<Codec<FourLeafCloverModifier>> FOUR_LEAF_CLOVER = SERIALIZERS.register("four_leaf_clover", () -> FourLeafCloverModifier.CODEC);
     public static final RegistryObject<Codec<HummingArtifactModifier>> HUMMING_ARTIFACT = SERIALIZERS.register("humming_artifact", () -> HummingArtifactModifier.CODEC);
+    public static final RegistryObject<Codec<EssenceThiefModifier>> ESSENCE_THIEF = SERIALIZERS.register("essence_thief", () -> EssenceThiefModifier.CODEC);
 }